[googlemaps] Blank for cleaning rudiments
