+++ types/googlemaps/reference/map.d.ts
@@ -1,4 +1,137 @@
 declare namespace google.maps {
+    interface MapHandlerMap {
+        /**
+         * This event is fired when the viewport bounds have changed.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.bounds_changed Maps JavaScript API}
+         * @see {@link Map#getBounds}
+         * @see {@link Map#fitBounds}
+         * @see {@link Map#panToBounds}
+         */
+        bounds_changed: [];
+
+        /**
+         * This event is fired when the map center property changes.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.center_changed Maps JavaScript API}
+         * @see {@link MapOptions#center}
+         * @see {@link Map#getCenter}
+         * @see {@link Map#setCenter}
+         */
+        center_changed: [];
+
+        /**
+         * This event is fired when the user clicks on the map.
+         * An ApiMouseEvent with properties for the clicked location is returned unless a place icon was clicked, in which case an IconMouseEvent with a placeid is returned.
+         * IconMouseEvent and ApiMouseEvent are identical, except that IconMouseEvent has the placeid field.
+         * The event can always be treated as an ApiMouseEvent when the placeid is not important.
+         * The click event is not fired if a marker or infowindow was clicked.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.click Maps JavaScript API}
+         */
+        click: [MouseEvent | IconMouseEvent];
+
+        /**
+         * This event is fired when the user double-clicks on the map. Note that the click event will also fire, right before this one.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.dblclick Maps JavaScript API}
+         */
+        dblclick: [MouseEvent];
+
+        /**
+         * This event is repeatedly fired while the user drags the map.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.drag Maps JavaScript API}
+         */
+        drag: [];
+
+        /**
+         * This event is fired when the user stops dragging the map.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.dragend Maps JavaScript API}
+         */
+        dragend: [];
+
+        /**
+         * This event is fired when the user starts dragging the map.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.dragstart Maps JavaScript API}
+         */
+        dragstart: [];
+
+        /**
+         * This event is fired when the map heading property changes.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.heading_changed Maps JavaScript API}
+         * @see {@link MapOptions#heading}
+         * @see {@link Map#getHeading}
+         * @see {@link Map#setHeading}
+         */
+        heading_changed: [];
+
+        /**
+         * This event is fired when the map becomes idle after panning or zooming.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.idle Maps JavaScript API}
+         */
+        idle: [];
+
+        /**
+         * This event is fired when the mapTypeId property changes.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.maptypeid_changed Maps JavaScript API}
+         * @see {@link MapOptions#mapTypeId}
+         * @see {@link Map#getMapTypeId}
+         * @see {@link Map#setMapTypeId}
+         */
+        maptypeid_changed: [];
+
+        /**
+         * This event is fired whenever the user's mouse moves over the map container.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.mousemove Maps JavaScript API}
+         */
+        mousemove: [MouseEvent];
+
+        /**
+         * This event is fired when the user's mouse exits the map container.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.mouseout Maps JavaScript API}
+         */
+        mouseout: [MouseEvent];
+
+        /**
+         * This event is fired when the user's mouse enters the map container.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.mouseover Maps JavaScript API}
+         */
+        mouseover: [MouseEvent];
+
+        /**
+         * This event is fired when the projection has changed.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.projection_changed Maps JavaScript API}
+         * @see {@link Map#getProjection}
+         */
+        projection_changed: [];
+
+        /**
+         * This event is fired when the DOM contextmenu event is fired on the map container.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.rightclick Maps JavaScript API}
+         */
+        rightclick: [MouseEvent];
+
+        /**
+         * This event is fired when the visible tiles have finished loading.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.tilesloaded Maps JavaScript API}
+         */
+        tilesloaded: [];
+
+        /**
+         * This event is fired when the map tilt property changes.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.tilt_changed Maps JavaScript API}
+         * @see {@link MapOptions#tilt}
+         * @see {@link Map#getTilt}
+         * @see {@link Map#setTilt}
+         */
+        tilt_changed: [];
+
+        /**
+         * This event is fired when the map zoom property changes.
+         * @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map.zoom_changed Maps JavaScript API}
+         * @see {@link MapOptions#zoom}
+         * @see {@link Map#getZoom}
+         * @see {@link Map#setZoom}
+         */
+        zoom_changed: [];
+    }
+
     /** @see {@link https://developers.google.com/maps/documentation/javascript/reference/map#Map Maps JavaScript API} */
     class Map<E extends Element = Element> extends MVCObject {
         /**
@@ -281,6 +414,8 @@ declare namespace google.maps {
          */
         setClickableIcons(clickable: boolean): void;
     }
+
+    type GestureHandlingOptions = 'cooperative' | 'greedy' | 'none' | 'auto';
 
     interface MapOptions {
         /**
